Separated the dashboard room list
